Update the quizz, debug of the quizz, debug slider, delete quizzResult.php, add some on style.css
While i have a bug on the quizz game, i choose to empty() the quizzDisplay then append the result.

For the slider, i added the preventDefault() on the <a> balises.

I added display:block; on the image of the slider.

I remove quizz1.js

transfer of some function from quizz.js to result.js